+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Features
 
+- Support for the installation of large canisters with chunked code of the Wasm module in `@dfinity/ic-management`.
 - Update most recent did files for ledgers (new optional field for initialization) and ckETH (new optional field in event).
 
 # 2024.03.05-1100Z